Refactor : PartyStatus 컴포넌트 생성
파티 상세 페이지에서 사용되었던 파티 상태 표시 부분을 컴포넌트로 분리하였습니다.
